requirements.txt
aliyun-python-sdk-core==2.13.4
aliyun-python-sdk-core-v3==2.8.6
aliyun-python-sdk-kms==2.9.0
#aliyunsdkcore==1.0.3
aiohttp==3.5.4
amqp==2.3.2
async-timeout==3.0.1
attrs==18.2.0
beautifulsoup4==4.6.0
billiard==3.5.0.3
celery==4.2.0
certifi==2018.4.16
chardet==3.0.4
idna==2.7
idna-ssl==1.1.0
kombu==4.2.1
motor==2.0.0
msgpack==0.5.6
multidict==4.5.2
Pillow==5.3.0
#pycrypto==2.6.1
#pycurl==7.43.0.2
pymongo==3.7.2
pytz==2018.4
#qcloudsms-py==0.1.3
#qiniu==7.2.2
redis==2.10.6
redis-py-cluster==1.3.5
requests==2.19.1
tornado==5.0.2
typing-extensions==3.7.4.1
urllib3==1.23
vine==1.1.4
xlrd==1.1.0
XlsxWriter==1.0.5
XML2Dict==0.2.2
yarl==1.3.0
#web3==5.4.0
#submail==0.4
ipfsapi==0.4.4
